moved some function from cvss.js to util.js
