+++ lib/cvss.js
@@ -1,5 +1,5 @@
 const definitions = require("./cvss_3_0.json");
-const { roundUp } = require("./util");
+const util = require("./util");
 
 /**
  * Creates a new CVSS object
@@ -7,52 +7,56 @@ const { roundUp } = require("./util");
  * @param {String} vector
  */
 function CVSS(vector) {
+
   /**
    * Retrieves an object of vector's metrics
-   *
+   * Calls a function from util.js
+   * 
    * @returns {Object} Abbreviations & Vector Value pair
    */
-  function getVectorObject() {
-    const vectorArray = vector.split("/");
-    const vectorObject = {};
-
-    for (const entry of vectorArray) {
-      const values = entry.split(":");
-      vectorObject[values[0]] = values[1];
-    }
-    return vectorObject;
+  function getVectorObject (){
+    return util.getVectorObject(vector);
   }
 
   /**
    * Retrieves an object of vector's metrics
+   * Calls a function from util.js
    *
    * @returns {Object} Abbreviations & Vectors Detailed Values
    */
   function getDetailedVectorObject() {
-    const vectorArray = vector.split("/");
-    const vectorObject = vectorArray.reduce(
-      (vectorObjectAccumulator, vectorItem, index) => {
-        const values = vectorItem.split(":");
-        const metrics = { ...vectorObjectAccumulator.metrics };
-        if (index) {
-          const vectorDef = findMetric(values[0]);
-          const detailedVectorObject = {
-            name: vectorDef.name,
-            abbr: vectorDef.abbr,
-            fullName: `${vectorDef.name} (${vectorDef.abbr})`,
-            value: vectorDef.metrics.find((def) => def.abbr === values[1]).name,
-            valueAbbr: values[1]
-          };
-          return Object.assign(vectorObjectAccumulator, {
-            metrics: Object.assign(metrics, { [values[0].trim()]: detailedVectorObject })
-          });
-        } else {
-          return Object.assign(vectorObjectAccumulator, { [values[0].trim()]: values[1] });
-        }
-      },
-      { metrics: {} }
-    );
-    return vectorObject;
+    return util.getDetailedVectorObject(vector);
+  }
+  
+  /**
+   * Calculates the Rating of the given vector
+   * Calls a function from util.js
+   * 
+   * @returns {String} returns one of the five possible ratings
+   */
+  function getRating() {
+    return util.getRating(getScore());
+  }
+
+  /**
+   * Checks if the given vector is valid
+   * Calls a function from util.js
+   * 
+   * @returns {boolean} valid = true ; not valid = false
+   */
+  const isVectorValid = function() {
+     return util.isVectorValid(vector);
+  }
+
+  /**
+   * Converts an object into a vectorstring
+   * Calls a function from util.js
+   * 
+   * @param {*} obj 
+   * @returns {String} returns the vectorstring
+   */
+  function parseVectorObjectToString(obj){
+    return util.parseVectorObjectToString(obj);
   }
 
   /**
@@ -72,33 +76,10 @@ function CVSS(vector) {
     const exploitability = calculateExploitability(vectorObject, scopeChanged);
 
     if (scopeChanged) {
-      return roundUp(Math.min(1.08 * (ISC + exploitability), 10), 1);
+      return util.roundUp(Math.min(1.08 * (ISC + exploitability), 10), 1);
     }
 
-    return roundUp(Math.min(ISC + exploitability, 10), 1);
-  }
-
-  /**
-   * Parses the vector to a rating
-   *
-   * @returns {String} Rating value
-   */
-  function getRating() {
-    const score = getScore();
-    let rating = "None";
-
-    if (score === 0) {
-      rating = "None";
-    } else if (score <= 3.9) {
-      rating = "Low";
-    } else if (score <= 6.9) {
-      rating = "Medium";
-    } else if (score <= 8.9) {
-      rating = "High";
-    } else {
-      rating = "Critical";
-    }
-    return rating;
+    return util.roundUp(Math.min(ISC + exploitability, 10), 1);
   }
 
   /**
@@ -111,9 +92,9 @@ function CVSS(vector) {
 
     const baseScore = getScore();
 
-    const eDef = findMetric("E");
-    const rlDef = findMetric("RL");
-    const rcDef = findMetric("RC");
+    const eDef = util.findMetric("E");
+    const rlDef = util.findMetric("RL");
+    const rcDef = util.findMetric("RC");
 
     const exploitCodeMaturity = eDef.metrics.find(
       (metric) => metric.abbr === (vectorObject.E ? vectorObject.E : "X")
@@ -125,7 +106,7 @@ function CVSS(vector) {
       (metric) => metric.abbr === (vectorObject.RC ? vectorObject.RC : "X")
     ).numerical;
 
-    return roundUp(baseScore * exploitCodeMaturity * remediationLevel * reportConfidence, 1);
+    return util.roundUp(baseScore * exploitCodeMaturity * remediationLevel * reportConfidence, 1);
   }
 
   /**
@@ -142,9 +123,9 @@ function CVSS(vector) {
 
     if (modifiedISC <= 0) return 0;
 
-    const eDef = findMetric("E");
-    const rlDef = findMetric("RL");
-    const rcDef = findMetric("RC");
+    const eDef = util.findMetric("E");
+    const rlDef = util.findMetric("RL");
+    const rcDef = util.findMetric("RC");
 
     const e = eDef.metrics.find((metric) => metric.abbr === vectorObject.E);
     const eValue = e ? e.numerical : 1;
@@ -154,13 +135,13 @@ function CVSS(vector) {
     const rcValue = rc ? rc.numerical : 1;
 
     if (!scopeChanged) {
-      return roundUp(
-        roundUp(Math.min(modifiedISC + modifiedExploitability, 10), 1) * eValue * rlValue * rcValue,
+      return util.roundUp(
+        util.roundUp(Math.min(modifiedISC + modifiedExploitability, 10), 1) * eValue * rlValue * rcValue,
         1
       );
     }
-    return roundUp(
-      roundUp(Math.min(1.08 * (modifiedISC + modifiedExploitability), 10), 1) *
+    return util.roundUp(
+      util.roundUp(Math.min(1.08 * (modifiedISC + modifiedExploitability), 10), 1) *
         eValue *
         rlValue *
         rcValue,
@@ -168,10 +149,11 @@ function CVSS(vector) {
     );
   }
 
+
   const calculateISCBase = function (vectorObject) {
-    const cDef = findMetric("C");
-    const iDef = findMetric("I");
-    const aDef = findMetric("A");
+    const cDef = util.findMetric("C");
+    const iDef = util.findMetric("I");
+    const aDef = util.findMetric("A");
 
     const cValue = cDef.metrics.find((metric) => metric.abbr === vectorObject.C).numerical;
     const iValue = iDef.metrics.find((metric) => metric.abbr === vectorObject.I).numerical;
@@ -187,12 +169,12 @@ function CVSS(vector) {
   };
 
   function calculateISCModifiedBase(vectorObject) {
-    const mcDef = findMetric("MC");
-    const miDef = findMetric("MI");
-    const maDef = findMetric("MA");
-    const crDef = findMetric("CR");
-    const irDef = findMetric("IR");
-    const arDef = findMetric("AR");
+    const mcDef = util.findMetric("MC");
+    const miDef = util.findMetric("MI");
+    const maDef = util.findMetric("MA");
+    const crDef = util.findMetric("CR");
+    const irDef = util.findMetric("IR");
+    const arDef = util.findMetric("AR");
 
     const mcValue = mcDef.metrics.find((metric) => metric.abbr === vectorObject.MC).numerical;
     const miValue = miDef.metrics.find((metric) => metric.abbr === vectorObject.MI).numerical;
@@ -208,10 +190,10 @@ function CVSS(vector) {
   }
 
   const calculateExploitability = function (vectorObject, scopeChanged) {
-    const avDef = findMetric("AV");
-    const acDef = findMetric("AC");
-    const prDef = findMetric("PR");
-    const uiDef = findMetric("UI");
+    const avDef = util.findMetric("AV");
+    const acDef = util.findMetric("AC");
+    const prDef = util.findMetric("PR");
+    const uiDef = util.findMetric("UI");
 
     const avValue = avDef.metrics.find((metric) => metric.abbr === vectorObject.AV).numerical;
     const acValue = acDef.metrics.find((metric) => metric.abbr === vectorObject.AC).numerical;
@@ -224,11 +206,10 @@ function CVSS(vector) {
   };
 
   const calculateModifiedExploitability = function (vectorObject, scopeChanged) {
-    const mavDef = findMetric("MAV");
-    const macDef = findMetric("MAC");
-    const mprDef = findMetric("MPR");
-    const muiDef = findMetric("MUI");
-
+    const mavDef = util.findMetric("MAV");
+    const macDef = util.findMetric("MAC");
+    const mprDef = util.findMetric("MPR");
+    const muiDef = util.findMetric("MUI");
     const mavValue = mavDef.metrics.find((metric) => metric.abbr === vectorObject.MAV).numerical;
     const macValue = macDef.metrics.find((metric) => metric.abbr === vectorObject.MAC).numerical;
     const muiValue = muiDef.metrics.find((metric) => metric.abbr === vectorObject.MUI).numerical;
@@ -238,114 +219,6 @@ function CVSS(vector) {
 
     return 8.22 * mavValue * macValue * mprValue * muiValue;
   };
-
-  const findMetric = function (abbr) {
-    return definitions.definitions.find((def) => def.abbr === abbr);
-  };
-
-  /**
-   * Checks whether the vector passed is valid
-   *
-   * @returns {Boolean} result with whether the vector is valid or not
-   */
-  const isVectorValid = function () {
-    /**
-     * This function is used to scan the definitions file and join all
-     * abbreviations in a format that RegExp understands.
-     *
-     * Exit example:
-     * ((((((((((AV:[NALP]|AC:[LH])|PR:[NLH])|UI:[NR])|S:[UC])|C:[NLW])|I:[NLW])|A:[NLW])|E:[XUPFH])|RL:[XOTWU])|RC:[XURC])
-     */
-    const expression = definitions.definitions.reduce((accumulator, currentValue, index) => {
-      const serializedAbbr = `${currentValue.abbr}:[${currentValue.metrics.reduce(
-        (accumulator2, currentValue2) => {
-          return accumulator2 + currentValue2.abbr;
-        },
-        ""
-      )}]`;
-      if (index !== 0) {
-        return `(${accumulator}|${serializedAbbr})`;
-      } else {
-        return serializedAbbr;
-      }
-    }, "");
-
-    const totalExpressionVector = new RegExp("^CVSS:3.0(/" + expression + ")+$");
-
-    //Checks if the vector is in valid format
-    if (!totalExpressionVector.test(vector)) {
-      return false;
-    }
-
-    /**
-     * Scans the definitions file and returns an array of each registered abbreviation
-     * with its possible values.
-     *
-     * Exit example:
-     * [/\/AV:[NALP]/g, /\/AC:[LH]/g, /\/PR:[NLH]/g, /\/UI:[NR]/g, /\/S:[UC]/g,]
-     *
-     * A / at the beginning serves for the algorithm not to confuse abbreviations as AC and C.
-     */
-    const allExpressions = definitions.definitions.map((currentValue) => {
-      return new RegExp(
-        `/${currentValue.abbr}:[${currentValue.metrics.reduce((accumulator2, currentValue2) => {
-          return accumulator2 + currentValue2.abbr;
-        }, "")}]`,
-        "g"
-      );
-    });
-
-    for (const regex of allExpressions) {
-      if ((vector.match(regex) || []).length > 1) {
-        return false;
-      }
-    }
-
-    const mandatoryParams = [
-      /\/AV:[NALP]/g,
-      /\/AC:[LH]/g,
-      /\/PR:[NLH]/g,
-      /\/UI:[NR]/g,
-      /\/S:[UC]/g,
-      /\/C:[NLH]/g,
-      /\/I:[NLH]/g,
-      /\/A:[NLH]/g
-    ];
-
-    //Checks whether all mandatory parameters are present in the vector
-    for (const regex of mandatoryParams) {
-      if ((vector.match(regex) || []).length < 1) {
-        return false;
-      }
-    }
-
-    return true;
-  };
-
-  /**
-   * This transforms an object in the format of getVectorObject()
-   * and parses it to a CVSS comaptible string
-   *
-   * @param {Object} obj
-   */
-  function parseVectorObjectToString(obj) {
-    if (typeof obj === "string") {
-      return obj;
-    }
-
-    let vectorString = `CVSS:${obj["CVSS"]}/`; 
-
-    for (const entry of definitions["definitions"]) {
-      const metric = entry["abbr"];
-      if (Object.prototype.hasOwnProperty.call(obj, metric)) {
-        vectorString += `${metric}:${obj[metric]}/`;
-      }
-    }
-
-    vectorString = vectorString.slice(0, -1);
-
-    return vectorString;
-  }
 
   vector = parseVectorObjectToString(vector);
